cmd: Split into separate files
Splitting these operations into dedicated files for each area will make
it easier to convert these areas into subcommands in an upcoming commit.
